+++ src/dgpost/transform/namespace.py
@@ -13,8 +13,7 @@ single namespace.
 """
 import pint
 
-from dgpost.transform.helpers import load_data
-from .helpers import columns_to_smiles
+from dgpost.utils.helpers import load_data, columns_to_smiles
 from collections import defaultdict
