+++ src/dgpost/utils/load.py
@@ -25,7 +25,7 @@ from uncertainties import ufloat_fromstr
 import re
 from typing import Union
 import logging
-from dgpost.transform.helpers import arrow_to_multiindex, set_units
+from dgpost.utils.helpers import arrow_to_multiindex
 
 logger = logging.getLogger(__name__)
