+++ src/dgpost/utils/plot.py
@@ -19,7 +19,7 @@ import uncertainties.unumpy as unp
 import logging
 import numpy as np
 
-from dgpost.transform.helpers import get_units, keys_in_df
+from dgpost.utils.helpers import get_units, keys_in_df
 
 logger = logging.getLogger(__name__)
